[BE] refactor(#7): User가 직접 UserDetails를 구현하도록 변경
* CustomUserDetails 삭제
