+++ backend/src/main/java/com/example/backend/auth/config/security/auth/JpaUserDetailsService.java
@@ -5,6 +5,7 @@ import com.example.backend.domain.define.user.User;
 import com.example.backend.domain.define.user.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
@@ -18,14 +19,13 @@ public class JpaUserDetailsService implements UserDetailsService {
     private UserRepository userRepository;
 
     @Override
-    public CustomUserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         // 사용자 인스턴스(Optional) 찾아서 반환
         User u = userRepository.findByEmail(username).orElseThrow(() -> {
             log.error(">>>> [ SECURITY ERROR: {} ]", ExceptionMessage.SECURITY_USER_NOT_FOUND.getText());
             throw new SecurityException(ExceptionMessage.SECURITY_USER_NOT_FOUND.getText());
         });
 
-        // 찾은 사용자를 UserDetails 타입으로 데코레이트해서 반환
-        return new CustomUserDetails(u);
+        return u;
     }
 }
